Add ClServer
